Translate MainActivity and DiagramFragment to Kotlin, main package, improve navigation, correct TreeUtil name

+++ app/src/main/java/app/familygem/main/NotesFragment.java
@@ -1,4 +1,4 @@
-package app.familygem.list;
+package app.familygem.main;
 
 import static app.familygem.Global.gc;
 
@@ -12,9 +12,9 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.SearchView;
-import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
@@ -23,30 +23,29 @@ import org.folg.gedcom.model.NoteContainer;
 import org.folg.gedcom.model.NoteRef;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import app.familygem.Memory;
-import app.familygem.Principal;
 import app.familygem.R;
 import app.familygem.U;
 import app.familygem.constant.Choice;
 import app.familygem.constant.Extra;
 import app.familygem.detail.NoteActivity;
-import app.familygem.util.TreeUtils;
+import app.familygem.util.TreeUtil;
 import app.familygem.util.Util;
 import app.familygem.visitor.FindStack;
 import app.familygem.visitor.NoteList;
 
-public class NotesFragment extends Fragment implements NotesAdapter.ItemClickListener {
+public class NotesFragment extends BaseFragment implements NotesAdapter.ItemClickListener {
 
-    private List<Note> allNotes;
+    private List<Note> allNotes = Collections.emptyList(); // Initialized to avoid crash in case Global.gc is null
     private NotesAdapter adapter;
 
     public static List<Note> getAllNotes(boolean sharedOnly) {
         // Shared notes
         List<Note> sharedNotes = gc.getNotes();
-        ArrayList<Note> noteList = new ArrayList<>();
-        noteList.addAll(sharedNotes);
+        ArrayList<Note> noteList = new ArrayList<>(sharedNotes);
         // Inline notes
         if (!sharedOnly) {
             NoteList noteVisitor = new NoteList();
@@ -58,33 +57,32 @@ public class NotesFragment extends Fragment implements NotesAdapter.ItemClickLis
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle bundle) {
+        super.onCreateView(inflater, container, bundle);
         View view = inflater.inflate(R.layout.recyclerview, container, false);
-        if (gc != null) {
-            RecyclerView recyclerView = view.findViewById(R.id.recycler_view);
-            recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
-            boolean sharedOnly = getActivity().getIntent().getBooleanExtra(Choice.NOTE, false);
-            allNotes = getAllNotes(sharedOnly);
-            adapter = new NotesAdapter(getContext(), allNotes, sharedOnly);
-            adapter.setClickListener(this);
-            recyclerView.setAdapter(adapter);
-            registerForContextMenu(recyclerView);
-            furnishToolbar();
-            view.findViewById(R.id.fab).setOnClickListener(v -> newNote(getContext(), null));
-        }
+        RecyclerView recyclerView = view.findViewById(R.id.recycler_view);
+        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+        boolean sharedOnly = getActivity().getIntent().getBooleanExtra(Choice.NOTE, false);
+        adapter = new NotesAdapter(getContext(), allNotes, sharedOnly);
+        adapter.setClickListener(this);
+        recyclerView.setAdapter(adapter);
+        registerForContextMenu(recyclerView);
+        view.findViewById(R.id.fab).setOnClickListener(v -> newNote(getContext(), null));
         return view;
     }
 
-    private void furnishToolbar() {
-        ((AppCompatActivity)getActivity()).getSupportActionBar().setTitle(allNotes.size() + " "
-                + Util.INSTANCE.caseString(allNotes.size() == 1 ? R.string.note : R.string.notes));
-        setHasOptionsMenu(allNotes.size() > 1);
+    @Override
+    public void showContent() {
+        boolean sharedOnly = getActivity().getIntent().getBooleanExtra(Choice.NOTE, false);
+        allNotes = getAllNotes(sharedOnly);
+        adapter.noteList = allNotes;
+        adapter.notifyDataSetChanged();
     }
 
-    // Leaving the activity without choosing a shared note resets the extra
     @Override
     public void onPause() {
         super.onPause();
-        getActivity().getIntent().removeExtra(Choice.NOTE);
+        // Resets the extra on leaving the activity without choosing a shared note
+        requireActivity().getIntent().removeExtra(Choice.NOTE);
     }
 
     @Override
@@ -110,45 +108,36 @@ public class NotesFragment extends Fragment implements NotesAdapter.ItemClickLis
 
     @Override
     public boolean onContextItemSelected(MenuItem item) {
-        if (item.getItemId() == 0) { // Delete
+        if (item.getGroupId() == 4 && item.getItemId() == 0) { // Delete
             Object[] leaders = U.deleteNote(adapter.selectedNote, null);
-            TreeUtils.INSTANCE.save(false, leaders);
-            refresh();
-            ((Principal)requireActivity()).furnishMenu();
-        } else {
-            return false;
+            TreeUtil.INSTANCE.save(false, leaders);
+            showContent();
+            ((MainActivity)requireActivity()).refreshInterface();
+            return true;
         }
-        return true;
+        return false;
     }
 
     @Override
-    public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
-        // Search inside notes
-        inflater.inflate(R.menu.search, menu);
-        final SearchView searchView = (SearchView)menu.findItem(R.id.search_item).getActionView();
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextChange(String query) {
-                adapter.getFilter().filter(query);
-                return true;
-            }
+    public void updateToolbar(ActionBar bar, Menu menu, MenuInflater inflater) {
+        bar.setTitle(allNotes.size() + " " + Util.INSTANCE.caseString(allNotes.size() == 1 ? R.string.note : R.string.notes));
+        if (allNotes.size() > 1) {  // Search inside notes
+            inflater.inflate(R.menu.search, menu);
+            final SearchView searchView = (SearchView)menu.findItem(R.id.search_item).getActionView();
+            searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+                @Override
+                public boolean onQueryTextChange(String query) {
+                    adapter.getFilter().filter(query);
+                    return true;
+                }
 
-            @Override
-            public boolean onQueryTextSubmit(String q) {
-                searchView.clearFocus();
-                return false;
-            }
-        });
-    }
-
-    /**
-     * Updates list and toolbar.
-     */
-    public void refresh() {
-        allNotes = getAllNotes(false);
-        adapter.noteList = allNotes;
-        adapter.notifyDataSetChanged();
-        furnishToolbar();
+                @Override
+                public boolean onQueryTextSubmit(String q) {
+                    searchView.clearFocus();
+                    return false;
+                }
+            });
+        }
     }
 
     /**
@@ -167,7 +156,7 @@ public class NotesFragment extends Fragment implements NotesAdapter.ItemClickLis
             noteRef.setRef(id);
             ((NoteContainer)container).addNoteRef(noteRef);
         }
-        TreeUtils.INSTANCE.save(true, note);
+        TreeUtil.INSTANCE.save(true, note);
         Memory.setLeader(note);
         context.startActivity(new Intent(context, NoteActivity.class));
     }
